PersolBullHornAppV2/requirements.txt
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
idna==3.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
traceback2==1.4.0
urllib3==1.26.8
beautifulsoup4
lxml
deepdiff
